+++ libs/Boy/Graphics.h
@@ -48,7 +48,7 @@ namespace Boy
 		 */
 		virtual void drawLine(float x0, float y0, float x1, float y1) = 0;
 		virtual void fillRect(int x0, int y0, int w, int h) = 0;
-		virtual void drawCircle(int x0, int y0, float r, int s) = 0;
+		virtual void drawCircle(float x0, float y0, float r, int s) = 0;
 		virtual void drawTriStrip(TriStrip *strip) = 0;
 		virtual void drawLineStrip(TriStrip *strip) = 0;
